feat: move common ECL operators to generic ECL evaluation
